requirements.txt
# core
pandas
pyyaml
rank-bm25
sentence-transformers
faiss-cpu
chromadb
python-docx
pypdf

# eval
ragas

# app
streamlit

# dev
pytest
ruff
